site settings done, functionality in tact, optimized requests
